Add BluetoothScoJob to manage startBluetoothSco retry logic
